[PF-405] Implement IAM auth for AWS MSK (Kafka) (#3244)
* add msk new dependency, add implementation for aws msk iam auth

* add refresh logic + unit tests

* add more unit tests

* add changelog entry, add unit test, update dependabot config

+++ rasa/core/iam_credentials_providers/credentials_provider_protocol.py
@@ -16,7 +16,7 @@ class TemporaryCredentials(BaseModel):
     """Dataclass storing temporary credentials."""
 
     auth_token: Optional[str] = None
-    expiration: Optional[int] = None
+    expiration: Optional[float] = None
     username: Optional[str] = None
     presigned_url: Optional[str] = None
